feat: Finish the linux parser in the first version
Finish the linux parser functions in the first version. The implementation is done with best effort but
is not garanteed to be absolutely correct. In the next step, I will recheck the implementation of
all functions in order to provide the correct functions for the use within the system monitor
functionallities.

No issue addressed