feat: Create a structure how to store and read the system and process information into the class datastructures
Create the structure how to hand over the system and process information to the visualization class written
with ncurses. The fixed data of the system and the fixed information of the processes (like "command" and "owner"
as well as the OS and kernel version) is stored as the member variables of the class datastructures. The
variable information is read out from the file system, every time when a method of the corresponding object
is called. The code compiles but I get an memory access error on runtime.

No issue addressed