feat: Correct the cpu utilization for processes and the system and implement the convertion from seconds to hh:mm:ss
Correct the calculation of the CPU utilization of the processes and the system as well update of the comparison
function for the processes based on the new calculation. Therefore, a new function is added to the linux_parser
file the returns the start time of a process that is needed for the calculation of the process CPU utilization.

No issue addressed