bugfix: Make the code build without warnings.
Some minor changes to the code so that the compiler does not produce errors. Therefore
the unused variables are marked with [maybe_unused] and the control loops are
rewritten such that they did not return anything within the control loop, instead
they return just after the else-statement.

No issue addressed